get rid of process_new_messages_loop, simplify the whole structure
